Revert "게시물작성, 비디오업로드 페이지 컴포넌트 분리"
