+++ src/components/Modal.js
@@ -7,43 +7,9 @@ import modal from "../assets/modal.svg";
 import defaultIMG from "../assets/myPage/profilePicture.png";
 
 const Modal = (props) => {
-  const { page, complete } = props;
+  const { profile, complete } = props;
 
-  if (page === 'board') {
-    return (
-      <ModalBackground onClick={props.closeModal} radius={props.radius}>
-        <ModalContainer width={props.width} height={props.height}>
-          <Menu>
-            <Image src={modal} width="25px" height="25px" margin="0 auto" />
-          </Menu>
-          <Menu
-            fontS={props.fontS}
-            padding={props.padding}
-            onClick={props.edit}
-          >
-            수정하기
-          </Menu>
-          <Menu
-            fontS={props.fontS}
-            padding={props.padding}
-            onClick={props.delete}
-          >
-            삭제하기
-          </Menu>
-          <Menu
-            fontS={props.fontS}
-            padding={props.padding}
-            opacity="0.5"
-            onClick={props.closemodal}
-          >
-            취소
-          </Menu>
-        </ModalContainer>
-      </ModalBackground>
-    )
-  }
-
-  if (page === 'profile') {
+  if (profile) {
     return (
       <ModalBackground onClick={props.closeModal}>
         <ModalContainer
