--- src/components/molecule/VideoWrite.js
@@ -1,13 +1,0 @@
-import React from "react";
-import BoardBody from "../BoardBody";
-
-const VideoUpload = (props) => {
-  
-  return (
-    <React.Fragment>
-      <BoardBody page='video'/>
-    </React.Fragment>
-  );
-};
-
-export default VideoUpload;
