+++ src/redux/modules/freeboard.js
@@ -1,7 +1,6 @@
 import { handleActions, createAction } from "redux-actions";
 import { apis } from "../../shared/apis";
 import produce from "immer";
-import { imageActions } from "./image";
 
 // initialState
 const initialState = {
@@ -59,7 +58,6 @@ export const addBoardDB =
       response && history.push(`/freeboardlist/${skiResort}`);
       history.push(`/freeboardlist/${skiResort}`);
       dispatch(addBoard(response.data));
-      dispatch(imageActions.setPreview(null));
     } catch (err) {
       console.log(`오류 발생!${err}`);
     }
@@ -100,7 +98,6 @@ export const updateBoardDB =
       const response = await apis.updateFreePost(postId, formdata);
       response && history.push(`/freeboarddetail/${skiResort}/${postId}`);
       dispatch(updateBoard(response.config.data));
-      dispatch(imageActions.setPreview(null));
     } catch (err) {
       console.log(`오류 발생!${err}`);
     }
